Refatoração da organização de arquivos.
