+++ package.json
@@ -10,8 +10,8 @@
     "test": "jest",
     "dev": "ts-node-dev --transpile-only --ignore-watch node_modules --respawn  src/server.ts",
     "typeorm": "typeorm-ts-node-commonjs",
-    "migration:up": "yarn typeorm migration:run -d ./src/migrations.config.ts",
-    "migration:down": "yarn typeorm migration:revert -d src/migrations.config.ts"
+    "migration:up": "yarn typeorm migration:run -d ./src/persistence/context/migrations.config.ts",
+    "migration:down": "yarn typeorm migration:revert -d ./src/persistence/context/migrations.config.ts"
   },
   "dependencies": {
     "babel-jest": "^29.3.1",
