--- src/Users/repositories/IAccountsRepositories.ts
@@ -1,10 +1,0 @@
-import { Account } from '../entities/Account'
-
-interface IAccountsRepositories {
-  create(): Promise<Account>
-  debit(account: Account, value: number): Promise<Account>
-  credit(account: Account, value: number): Promise<Account>
-  getBalance(username: string): Promise<Number>
-}
-
-export { IAccountsRepositories }
