--- src/Users/repositories/IUsersRepositories.ts
@@ -1,8 +1,0 @@
-import { User } from '../entities/User'
-
-interface IUsersRepositories {
-  save(user: User): Promise<User>
-  findByUsername(username: string): Promise<User | undefined>
-}
-
-export { IUsersRepositories }
