+++ src/domain/models/User.ts
