+++ src/routes/transactions.routes.ts
@@ -2,15 +2,15 @@ import { Request, Response, Router } from 'express'
 import { ensureAuthenticated } from '../middlewares/ensureAutenticated'
 import { ensureGetTransactiondata } from '../middlewares/ensureGetTransactiondata'
 import { ensureMakeTransferData } from '../middlewares/ensureMakeTransferData'
-import { CreateTransactionController } from '../Transactions/useCases/createTransaction/CreateTransactionController'
-import { GetTransactionsController } from '../Transactions/useCases/getTransaction/GetTransactionsController'
-import { UsersRepositories } from '../Users/repositories/implementations/UsersRepositories'
-import { AccountsRepositories } from '../Users/repositories/implementations/AccountsRepositories'
-import { createConnection } from '../databaseConnection'
-import { TransactionsRepositories } from '../Transactions/repositories/implementations/TransactionsRepositories'
-import { User } from '../Users/entities/User'
-import { Account } from '../Users/entities/Account'
-import { Transaction } from '../Transactions/entities/Transaction'
+import { CreateTransactionController } from '../controllers/transactionsControllers/createTransaction/CreateTransactionController'
+import { GetTransactionsController } from '../controllers/transactionsControllers/getTransaction/GetTransactionsController'
+import { AccountsRepositories } from '../persistence/repositories/AccountsRepositories'
+import { createConnection } from '../persistence/context/databaseConnection'
+import { Account } from '../domain/models/Account'
+import { Transaction } from '../domain/models/Transaction'
+import { UsersRepositories } from '../persistence/repositories/UsersRepositories'
+import { User } from '../domain/models/User'
+import { TransactionsRepositories } from '../persistence/repositories/TransactionsRepositories'
 
 const transactionsRoutes = Router()
